:truck: Move class to Geoportal model
